+++ src/pages/account/AccountPage.scss
@@ -1,0 +1,5 @@
+.account-page {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
