--- src/pages/account/form/Form.css
@@ -1,14 +1,0 @@
-.form {
-  display: flex;
-  justify-content: center;
-  margin: 0 auto;
-  /* background-color:#f3f3f3; */
-}
-
-.form__container {
-  padding: 32px 64px;
-  display: flex;
-  flex-direction: column;
-  align-items: baseline;
-  gap: 30px;
-}
