+++ src/pages/account/form/Form.scss
@@ -1,0 +1,12 @@
+.form {
+  display: flex;
+  justify-content: center;
+  margin: 0 auto;
+
+  &__container {
+    display: flex;
+    flex-direction: column;
+    align-items: baseline;
+    gap: 30px;
+  }
+}
