--- src/pages/account/input/Input.css
@@ -1,12 +1,0 @@
-.input-container {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.input {
-  height: 40px;
-  width: 280px;
-  border: solid 1px gray;
-  cursor: pointer;
-}
