+++ src/pages/account/input/Input.scss
@@ -1,0 +1,30 @@
+@import '../../../styles/variables.scss';
+
+.input-container {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+
+  label {
+    font-size: 1.6rem;
+    font-weight: 700;
+  }
+
+  .input {
+    height: 56px;
+    width: 400px;
+    border: solid 1px $bg-color-gray;
+    padding-left: 2px;
+    font-size: 1.6rem;
+    cursor: pointer;
+  }
+
+  .input[type='submit'] {
+    font-family: 'Bebas Neue';
+    font-weight: 700;
+    font-size: 2.4rem;
+    letter-spacing: 1px;
+    color: white;
+    background-color: $color-primary-black;
+  }
+}
