+++ src/pages/account/input/Input.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import './Input.css';
+import './Input.scss';
 
 interface IpropsInput {
-  name: string;
+  name?: string;
   type: string;
   placeholder?: string;
   value?: string;
